finalized refactoring
