+++ src/services/tasks/copy-and-paste/documentdb/documentDbDocumentReader.ts
@@ -1,0 +1,56 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { type Document, type WithId } from 'mongodb';
+import { ClustersClient } from '../../../../documentdb/ClustersClient';
+import { type DocumentDetails, type DocumentReader } from '../documentInterfaces';
+
+/**
+ * DocumentDB-specific implementation of DocumentReader.
+ */
+export class DocumentDbDocumentReader implements DocumentReader {
+    /**
+     * Streams documents from a DocumentDB collection.
+     *
+     * @param connectionId Connection identifier to get the DocumentDB client
+     * @param databaseName Name of the database
+     * @param collectionName Name of the collection
+     * @returns AsyncIterable of document details
+     */
+    async *streamDocuments(
+        connectionId: string,
+        databaseName: string,
+        collectionName: string,
+    ): AsyncIterable<DocumentDetails> {
+        const client = await ClustersClient.getClient(connectionId);
+
+        const documentStream = client.streamDocuments(databaseName, collectionName, new AbortController().signal);
+        for await (const document of documentStream) {
+            yield {
+                id: (document as WithId<Document>)._id,
+                documentContent: document,
+            };
+        }
+    }
+
+    /**
+     * Counts the total number of documents in the DocumentDB collection.
+     *
+     * @param connectionId Connection identifier to get the DocumentDB client
+     * @param databaseName Name of the database
+     * @param collectionName Name of the collection,
+     * @param filter Optional filter to apply to the count operation (default is '{}')
+     * @returns Promise resolving to the document count
+     */
+    async countDocuments(
+        connectionId: string,
+        databaseName: string,
+        collectionName: string,
+        filter: string = '{}',
+    ): Promise<number> {
+        const client = await ClustersClient.getClient(connectionId);
+        return await client.countDocuments(databaseName, collectionName, filter);
+    }
+}
